LPS-65760 Move classes around: commands are not public, they're part of our internal algorithm
