+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/internal/command/impl/DeleteExportProcessCommand.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.tools.data.partitioning.sql.builder.internal.command;
+package com.liferay.portal.tools.data.partitioning.sql.builder.internal.command.impl;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.command.BaseExportProcessCommand;
 import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.DBExporter;
 import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.context.ExportContext;
+import com.liferay.portal.tools.data.partitioning.sql.builder.internal.command.BaseExportProcessCommand;
 
 import java.io.OutputStream;
 
@@ -25,10 +25,9 @@ import java.util.List;
 /**
  * @author Manuel de la Peña
  */
-public class InsertPartitionedExportProcessCommand
-	extends BaseExportProcessCommand {
+public class DeleteExportProcessCommand extends BaseExportProcessCommand {
 
-	public InsertPartitionedExportProcessCommand(
+	public DeleteExportProcessCommand(
 		long companyId, DBExporter dbExporter, List<String> tableNames,
 		ExportContext exportContext) {
 
@@ -36,19 +35,17 @@ public class InsertPartitionedExportProcessCommand
 	}
 
 	protected String getOutputFileName() {
-		return
-			exportContext.getSchemaName() + "-" + companyId +
-				"-partitioned.sql";
+		return exportContext.getSchemaName() + "-" + companyId + "-delete.sql";
 	}
 
 	protected String getOutputFileName(String tableName) {
 		return
 			exportContext.getSchemaName() + "-" + companyId + "-table-" +
-				tableName + ".sql";
+				tableName + "-delete.sql";
 	}
 
 	protected void write(String tableName, OutputStream outputStream) {
-		dbExporter.write(companyId, tableName, outputStream);
+		dbExporter.writeDelete(companyId, tableName, outputStream);
 	}
 
 }
